pyproject.toml
[project]
name = "financial-ai-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anywidget>=0.9.21",
    "arch>=8.0.0",
    "backtesting>=0.6.5",
    "bitsandbytes>=0.48.2",
    "chromadb>=1.3.5",
    "colorama>=0.4.6",
    "datasets>=4.4.1",
    "fastapi>=0.124.0",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.1",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langgraph",
    "matplotlib>=3.10.7",
    "nbstripout>=0.8.2",
    "networkx>=3.6.1",
    "neuralforecast",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "peft>=0.18.0",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.4.1",
    "pytest>=9.0.2",
    "python-dotenv",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.45",
    "statsmodels>=0.14.6",
    "streamlit>=1.52.1",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.14",
    "textblob>=0.19.0",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "trl>=0.26.0",
    "uvicorn>=0.38.0",
    "vectorbt",
    "xgboost",
    "yfinance>=0.2.66",
]
